requirements.txt
alembic==0.8.8
Flask==0.11.1
Flask_Bcrypt==0.7.1
Flask_Login==0.4.0
Flask_Migrate==2.0.0
Flask_Script==2.0.5
Flask_SQLAlchemy==2.1
Flask_WTF==0.13.1
SQLAlchemy==1.1.3
WTForms==2.1
WTForms_SQLAlchemy==0.1
unicode_slugify==0.1.3
